Include Math and Date tools
git-svn-id: svn://192.168.1.132/autoStock/autoStock@66 33c8eb2e-3fb0-4f7a-9790-3894f54e5fb4

+++ src/com/autoStock/trading/platform/ib/IbExchangeWrapper.java
@@ -181,7 +181,7 @@ public class IbExchangeWrapper implements EWrapper {
 		if (date.contains("finished")){
 			((RequestHistoricalData)RequestManager.getRequestHolder(requestId).caller).finished();
 		}else{
-			((RequestHistoricalData)RequestManager.getRequestHolder(requestId).caller).addResult(new ExResultHistoricalData(). new ExResultRowHistoricalData(date, (float)close));
+			((RequestHistoricalData)RequestManager.getRequestHolder(requestId).caller).addResult(new ExResultHistoricalData(). new ExResultRowHistoricalData(Long.valueOf(date), close));
 		}
 	}
